requirements.txt
websockets==11.0.3
httpx==0.25.2
python-telegram-bot==20.7
loguru==0.7.0
asyncio==3.4.3
python-dotenv==1.0.0
aiohttp==3.9.0
pandas==2.0.3
